Create AD_Group_Computer_Audit
I was asked to audit some computers and groups in AD.  The catch was they didn't want just a straight dump of users.  They wanted all of the members along with which groups gave them access to an environment.

The problem I ran into with existing scripts on the internet (that I could find) is that they didn't support multiple domains very well.  All of them would break the moment one domain group contained a nested security group from another domain (foreign security principal?).  They would either halt there or not do any recursive lookups beyond that base level.

This was way over my head to, write so I commissioned the help from one of our coders (I got his permission to post this out here to get feedback and look for improvements).  

After several revisions he produced this gem.  It works beautifully in our environment and gives us something to build off of and refine the output to our liking.  So, I thought it would be worth sharing with others who may be in a similar situation looking for a solution.